+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -15,6 +15,9 @@
 * Message settlement methods (`complete`, `abandon`, `defer` and `dead_letter`)
 and methods that use amqp management link for request like `schedule_messages`, `received_deferred_messages`, etc.
 now raise more concrete exception other than `MessageSettleFailed` and `ServiceBusError`.
+* Exceptions `MessageSendFailed`, `MessageSettleFailed` and `MessageLockExpired`
+ now inherit from `azure.servicebus.exceptions.MessageError`.
+* Removed Exception `ServiceBusResourceNotFound` as `azure.core.exceptions.ResourceNotFoundError` is now raised when a Service Bus resource does not exist.
 
 **BugFixes**